+++ src/local_pooling_gpu.cu
@@ -188,7 +188,7 @@ at::Tensor LocalPoolingBackwardGPU(
               grad_in_feat.template data_ptr<scalar_t>(), in_feat.size(0),
               grad_out_feat.template data_ptr<scalar_t>(),
               grad_out_feat.size(0), num_nonzero.data_ptr<int>(),
-              in_feat.size(1), stream);
+              in_feat.size(1));
         });
   } else {
     AT_DISPATCH_FLOATING_TYPES(