+++ src/pooling_max_kernel.cu
@@ -42,10 +42,10 @@
 #include "pooling_max_kernel.cuh"
 #include "utils.hpp"
 
-template <typename Dtype, typename Itype>
-__global__ void set_gradient(const int n, const Dtype *d_grad_out,
-                             Dtype *d_grad_in, const Itype *in_index,
-                             const int unused_key) {
+template <typename Dtype, typename MaskItype>
+__global__ void set_gradient(const size_t n, const Dtype *d_grad_out,
+                             Dtype *d_grad_in, const MaskItype *in_index,
+                             const Dtype unused_key) {
   CUDA_KERNEL_LOOP(index, n) {
     auto const queried_index = in_index[index];
     if (queried_index != unused_key)
@@ -53,26 +53,30 @@ __global__ void set_gradient(const int n, const Dtype *d_grad_out,
   }
 }
 
-template <typename Dtype, typename Itype>
-__global__ void max_pool(const int N, const int out_nrows, const int nchannel,
-                         const int nmap, const Dtype *d_in_feat,
-                         Dtype *d_out_feat, int32_t *d_max_index,
-                         const Itype *d_in_map, const Itype *d_out_map,
-                         const Itype *d_in_index_min) {
+template <typename Dtype, typename MaskItype, typename MapItype>
+__global__ void max_pool_reduced_pointer_kernel_gpu(
+    const int N, const int out_nrows, const int nchannel, const int nmap,
+    const Dtype *__restrict__ d_in_feat,    //
+    Dtype *__restrict__ d_out_feat,         //
+    MaskItype *__restrict__ d_max_index,    //
+    const MapItype *__restrict__ d_in_map,  //
+    const MapItype *__restrict__ d_out_map, //
+    const MapItype *__restrict__ d_in_index_min) {
   // N == nmap * nchannel
   CUDA_KERNEL_LOOP(index, N) {
     int const nrow = index / nchannel;
     int const ch = index % nchannel;
 
-    Itype const out_map_row = d_out_map[nrow];
-    Itype const in_index = d_in_index_min[nrow];
-    Itype num_in_feat;
+    MapItype const out_map_row = d_out_map[nrow];
+    MapItype const in_index = d_in_index_min[nrow];
+    MapItype num_in_feat;
+
     if (nrow == out_nrows - 1)
       num_in_feat = nmap - in_index;
     else
       num_in_feat = d_in_index_min[nrow + 1] - in_index;
     // It is guaranteed to have at least one input per output
-    Itype curr_index, max_index = d_in_map[in_index] * nchannel + ch;
+    MapItype curr_index, max_index = d_in_map[in_index] * nchannel + ch;
     Dtype curr_val, max_val = d_in_feat[max_index];
     for (int curr_iter = 0; curr_iter < num_in_feat; ++curr_iter) {
       curr_index = d_in_map[in_index + curr_iter] * nchannel + ch;
@@ -82,7 +86,7 @@ __global__ void max_pool(const int N, const int out_nrows, const int nchannel,
         max_index = curr_index;
       }
     }
-    Itype const out_ind = out_map_row * nchannel + ch;
+    MapItype const out_ind = out_map_row * nchannel + ch;
     // TODO thrust::reduce_by_key results in erroneous results at the end for
     // very large array
     if (out_map_row < out_nrows) {
@@ -97,8 +101,9 @@ __global__ void max_pool(const int N, const int out_nrows, const int nchannel,
 // index
 template <typename Dtype, typename Itype>
 __global__ void copy_sorted(const int n, const int nrows, const int nchannel,
-                            const Dtype *in_feat, const Itype *in_index,
-                            Dtype *out_feat) {
+                            const Dtype *__restrict__ in_feat,
+                            const Itype *__restrict__ in_index,
+                            Dtype *__restrict__ out_feat) {
   int nrow, ch;
   CUDA_KERNEL_LOOP(index, n) {
     nrow = index / nchannel;
@@ -112,7 +117,7 @@ namespace minkowski {
 namespace detail {
 
 template <typename Dtype>
-__global__ void fill(const int n, Dtype *dst, Dtype const val) {
+__global__ void fill(const size_t n, Dtype *dst, Dtype const val) {
   auto const tx = threadIdx.x;
   auto const bx = blockIdx.x;
   auto const x = blockDim.x * bx + tx;
@@ -122,52 +127,39 @@ __global__ void fill(const int n, Dtype *dst, Dtype const val) {
 
 } // namespace detail
 
-template <typename Dtype, typename Itype, typename ByteAllocator>
-void MaxPoolingForwardKernelGPU(
-    const Dtype *d_in_feat, Dtype *d_out_feat, int out_nrows, int *d_max_index,
-    int nchannel, gpu_kernel_map<Itype, ByteAllocator> const &kernel_map,
-    ByteAllocator &allocator, cudaStream_t stream) {
+template <typename Dtype, typename MaskItype, typename MapItype>
+void max_pool_forward_pointer_kernel_gpu(
+    MapItype *d_in_map,     // this will be sorted
+    MapItype *d_out_map,    // this will be sorted
+    size_t const nmap,      // map size
+    Dtype const *d_in_feat, //
+    Dtype *d_out_feat,      //
+    size_t const out_nrows, //
+    size_t const nchannel,  //
+    MaskItype *d_max_index, //
+    bool const is_sorted    //
+) {
 
-  size_t nmap = kernel_map.size();
-  size_t scratch_bytes = 5 * (kernel_map.size() + 1) * sizeof(Itype);
-  Itype *d_scr = (Itype *)allocator.allocate(scratch_bytes);
-  Itype *d_in_map = d_scr, *d_out_map = d_scr + nmap + 1;
-  Itype *d_curr_in_map = d_in_map;
-  Itype *d_curr_out_map = d_out_map;
+  MapItype *d_scr = (MapItype *)c10::cuda::CUDACachingAllocator::raw_alloc(
+      3 * (nmap + 1) * sizeof(MapItype));
 
-#ifdef DEBUG
-  cudaMemset(d_scr, 0, scratch_bytes);
-  std::cout << "out_nrows: " << out_nrows << ", nmap: " << nmap << "\n";
-#endif
+  MapItype *d_index = d_scr;                          // sequence
+  MapItype *d_in_map_min = d_scr + 1 * nmap + 1;      // reduced min output maps
+  MapItype *d_reduced_out_map = d_scr + 2 * nmap + 2; // reduced output maps
 
-  for (auto k = kernel_map.key_cbegin(); k != kernel_map.key_cend(); ++k) {
-    auto kernel_index = k->first;
-    size_t curr_size = kernel_map.in_maps.size(kernel_index);
-    CUDA_CHECK(cudaMemcpyAsync(
-        d_curr_in_map, kernel_map.in_maps.begin(kernel_index),
-        curr_size * sizeof(int), cudaMemcpyDeviceToDevice, stream));
-    CUDA_CHECK(cudaMemcpyAsync(
-        d_curr_out_map, kernel_map.out_maps.begin(kernel_index),
-        curr_size * sizeof(int), cudaMemcpyDeviceToDevice, stream));
-    d_curr_in_map += curr_size;
-    d_curr_out_map += curr_size;
-  }
-  CUDA_CHECK(cudaStreamSynchronize(stream));
-
-  // First, sort d_out_map and d_in_map with the d_out_map so that in_feat are
-  // placed adjacent according to out_map
-  thrust::sort_by_key(thrust::device, d_out_map, d_out_map + nmap, d_in_map);
-
-  // Second, create number of in_feat per out, and starting index
-  Itype *d_index, *d_in_map_min, *d_reduced_out_map;
-  d_index = d_scr + 2 * nmap + 2;
-  d_in_map_min = d_scr + 3 * nmap + 3;
-  d_reduced_out_map = d_scr + 4 * nmap + 4;
-
+  // create number of in_feat per out, and starting index
   thrust::sequence(thrust::device, d_index, d_index + nmap);
 
-  thrust::equal_to<Itype> equal_pred;
-  thrust::minimum<Itype> min_op;
+  ////////////////////////////////
+  // Reduction
+  ////////////////////////////////
+  // sort d_out_map and d_in_map with the d_out_map so that in_feat are
+  // placed adjacent according to out_map
+  if (!is_sorted)
+    thrust::sort_by_key(thrust::device, d_out_map, d_out_map + nmap, d_in_map);
+
+  thrust::equal_to<MapItype> equal_pred;
+  thrust::minimum<MapItype> min_op;
 
   auto reduction_pair =
       thrust::reduce_by_key(thrust::device,    // execution policy
@@ -179,15 +171,16 @@ void MaxPoolingForwardKernelGPU(
                             equal_pred,        // binary pred
                             min_op);           // binary op
   CUDA_CHECK(cudaStreamSynchronize(0));
+
   size_t num_unique_out_map = reduction_pair.first - d_reduced_out_map;
 
 #ifdef DEBUG
   std::cout << "num_unique_out_map: " << num_unique_out_map << "\n";
-  Itype *p_scr = (Itype *)std::malloc((nmap + 1) * 2 * sizeof(Itype));
-  CUDA_CHECK(cudaMemcpy(p_scr, d_in_map_min, (nmap + 1) * 2 * sizeof(Itype),
+  MapItype *p_scr = (MapItype *)std::malloc((nmap + 1) * 2 * sizeof(MapItype));
+  CUDA_CHECK(cudaMemcpy(p_scr, d_in_map_min, (nmap + 1) * 2 * sizeof(MapItype),
                         cudaMemcpyDeviceToHost));
-  Itype step = std::max<Itype>(num_unique_out_map / 100, 1);
-  Itype i = 0;
+  MapItype step = std::max<MapItype>(num_unique_out_map / 100, 1);
+  MapItype i = 0;
   for (; i < num_unique_out_map;) {
     std::cout << i;
     std::cout << " in_map_min: " << p_scr[i]
@@ -203,6 +196,7 @@ void MaxPoolingForwardKernelGPU(
   std::free(p_scr);
   std::cout << "done printing\n";
 #endif
+
   if (num_unique_out_map > out_nrows)
     throw std::invalid_argument(
         Formatter() << "Invalid number of out nrows: " << out_nrows
@@ -210,88 +204,216 @@ void MaxPoolingForwardKernelGPU(
 
   // fill it with unused key
   detail::fill<<<GET_BLOCKS(out_nrows * nchannel, CUDA_NUM_THREADS),
-                 CUDA_NUM_THREADS, 0, stream>>>(
-      out_nrows * nchannel, d_max_index, std::numeric_limits<int>::max());
-  CUDA_CHECK(cudaStreamSynchronize(stream));
+                 CUDA_NUM_THREADS>>>(out_nrows * nchannel, d_max_index,
+                                     std::numeric_limits<MaskItype>::max());
+
 #ifdef DEBUG
+  // CUDA_CHECK(cudaStreamSynchronize(stream));
   std::cout << "filled\n";
 #endif
 
   // Finally, use the max kernel to map all in_feats with the same out key to
   // out_feats Also, create out max_index for gradient
-  max_pool<Dtype, Itype>
+  max_pool_reduced_pointer_kernel_gpu<Dtype, MaskItype, MapItype>
       <<<GET_BLOCKS(num_unique_out_map * nchannel, CUDA_NUM_THREADS),
-         CUDA_NUM_THREADS, 0, stream>>>(nchannel * num_unique_out_map, // N
-                                        num_unique_out_map, nchannel, nmap,
-                                        d_in_feat, d_out_feat,
-                                        d_max_index, // Out indices for backward
-                                        d_in_map,    // in index
-                                        d_reduced_out_map, d_in_map_min);
+         CUDA_NUM_THREADS>>>(nchannel * num_unique_out_map, // N
+                             num_unique_out_map, nchannel, nmap, d_in_feat,
+                             d_out_feat,
+                             d_max_index, // Out indices for backward
+                             d_in_map,    // in index
+                             d_reduced_out_map, d_in_map_min);
 
   // cudaFree(d_in_map);
   // cudaFree(d_index);
   CUDA_CHECK(cudaGetLastError());
+  // CUDA_CHECK(cudaStreamSynchronize(stream));
+
+  c10::cuda::CUDACachingAllocator::raw_delete((void *)d_scr);
+}
+
+template <typename Dtype, typename Itype, typename ByteAllocator>
+void MaxPoolingForwardKernelGPU(
+    const Dtype *d_in_feat, Dtype *d_out_feat, size_t const out_nrows,
+    int *d_max_index, size_t const nchannel,
+    gpu_kernel_map<Itype, ByteAllocator> const &kernel_map,
+    ByteAllocator &allocator, cudaStream_t stream) {
+
+  size_t nmap = kernel_map.size();
+  size_t scratch_bytes = 2 * (nmap + 1) * sizeof(Itype);
+  Itype *d_scr = (Itype *)allocator.allocate(scratch_bytes);
+
+  Itype *d_in_map = d_scr;             // all input kernel maps
+  Itype *d_out_map = d_scr + nmap + 1; // all output kernel maps
+
+  ////////////////////////////////
+  // Initialize data
+  ////////////////////////////////
+#ifdef DEBUG
+  cudaMemset(d_scr, 0, scratch_bytes);
+  std::cout << "out_nrows: " << out_nrows << ", nmap: " << nmap << "\n";
+#endif
+  Itype *d_curr_in_map = d_in_map;
+  Itype *d_curr_out_map = d_out_map;
+
+  for (auto k = kernel_map.key_cbegin(); k != kernel_map.key_cend(); ++k) {
+    auto kernel_index = k->first;
+    size_t curr_size = kernel_map.in_maps.size(kernel_index);
+    CUDA_CHECK(cudaMemcpyAsync(
+        d_curr_in_map, kernel_map.in_maps.begin(kernel_index),
+        curr_size * sizeof(int), cudaMemcpyDeviceToDevice, stream));
+    CUDA_CHECK(cudaMemcpyAsync(
+        d_curr_out_map, kernel_map.out_maps.begin(kernel_index),
+        curr_size * sizeof(int), cudaMemcpyDeviceToDevice, stream));
+    d_curr_in_map += curr_size;
+    d_curr_out_map += curr_size;
+  }
   CUDA_CHECK(cudaStreamSynchronize(stream));
+
+  //
+  max_pool_forward_pointer_kernel_gpu<Dtype, int32_t, Itype>(
+      d_in_map, d_out_map, nmap, d_in_feat, d_out_feat, out_nrows, nchannel,
+      d_max_index, false);
+
   allocator.deallocate((char *)d_scr, scratch_bytes);
 }
 
 // default_allocator
 template void
 MaxPoolingForwardKernelGPU<float, uint32_t, detail::default_allocator<char>>(
-    const float *d_in_feat, float *d_out_feat, int out_nrows,
-    int32_t *d_max_index, int nchannel,
+    const float *d_in_feat, float *d_out_feat, size_t const out_nrows,
+    int32_t *d_max_index, size_t const nchannel,
     gpu_kernel_map<uint32_t, detail::default_allocator<char>> const &kernel_map,
     detail::default_allocator<char> &allocator, cudaStream_t stream);
 
 template void
 MaxPoolingForwardKernelGPU<double, uint32_t, detail::default_allocator<char>>(
-    const double *d_in_feat, double *d_out_feat, int out_nrows,
-    int32_t *d_max_index, int nchannel,
+    const double *d_in_feat, double *d_out_feat, size_t const out_nrows,
+    int32_t *d_max_index, size_t const nchannel,
     gpu_kernel_map<uint32_t, detail::default_allocator<char>> const &kernel_map,
     detail::default_allocator<char> &allocator, cudaStream_t stream);
 
 // c10_allocator
 template void
 MaxPoolingForwardKernelGPU<float, uint32_t, detail::c10_allocator<char>>(
-    const float *d_in_feat, float *d_out_feat, int out_nrows,
-    int32_t *d_max_index, int nchannel,
+    const float *d_in_feat, float *d_out_feat, size_t const out_nrows,
+    int32_t *d_max_index, size_t const nchannel,
     gpu_kernel_map<uint32_t, detail::c10_allocator<char>> const &kernel_map,
     detail::c10_allocator<char> &allocator, cudaStream_t stream);
 
 template void
 MaxPoolingForwardKernelGPU<double, uint32_t, detail::c10_allocator<char>>(
-    const double *d_in_feat, double *d_out_feat, int out_nrows,
-    int32_t *d_max_index, int nchannel,
+    const double *d_in_feat, double *d_out_feat, size_t const out_nrows,
+    int32_t *d_max_index, size_t const nchannel,
     gpu_kernel_map<uint32_t, detail::c10_allocator<char>> const &kernel_map,
     detail::c10_allocator<char> &allocator, cudaStream_t stream);
 
-template <typename Dtype>
-void MaxPoolingBackwardKernelGPU(Dtype *d_grad_in_feat, int in_nrows,
-                                 const Dtype *d_grad_out_feat, int out_nrows,
-                                 const int32_t *d_max_index, int nchannel,
-                                 cudaStream_t stream) {
-  int num_kernels = out_nrows * nchannel;
+template <typename Dtype, typename MaskItype>
+void MaxPoolingBackwardKernelGPU(Dtype *d_grad_in_feat, size_t const in_nrows,
+                                 const Dtype *d_grad_out_feat,
+                                 size_t const out_nrows,
+                                 const MaskItype *d_max_index,
+                                 size_t const nchannel) {
+  size_t const num_kernels = out_nrows * nchannel;
   // Assume that gradients for input feature are all set to zero
   LOG_DEBUG("MaxPool Backward GPU with #", num_kernels, out_nrows, nchannel);
-  set_gradient<Dtype>
-      <<<GET_BLOCKS(num_kernels, CUDA_NUM_THREADS), CUDA_NUM_THREADS, 0,
-         stream>>>(num_kernels, d_grad_out_feat, d_grad_in_feat, d_max_index,
-                   std::numeric_limits<int>::max());
+  set_gradient<Dtype, MaskItype>
+      <<<GET_BLOCKS(num_kernels, CUDA_NUM_THREADS), CUDA_NUM_THREADS>>>(
+          num_kernels, d_grad_out_feat, d_grad_in_feat, d_max_index,
+          std::numeric_limits<MaskItype>::max());
 
-  CUDA_CHECK(cudaStreamSynchronize(stream));
+  // CUDA_CHECK(cudaStreamSynchronize(stream));
 }
 
-template void
-MaxPoolingBackwardKernelGPU<float>(float *d_grad_in_feat, int in_nrows,
-                                   const float *d_grad_out_feat, int out_nrows,
-                                   const int32_t *d_max_index, int nchannel,
-                                   cudaStream_t stream);
+template void max_pool_forward_pointer_kernel_gpu<float, int32_t, uint32_t>(
+    uint32_t *d_in_map,     //
+    uint32_t *d_out_map,    //
+    size_t const nmap,      //
+    float const *d_in_feat, //
+    float *d_out_feat,      //
+    size_t const out_nrows, //
+    size_t const nchannel,  //
+    int32_t *d_max_index,   //
+    bool const is_sorted    //
+);
 
-template void
-MaxPoolingBackwardKernelGPU<double>(double *d_grad_in_feat, int in_nrows,
-                                    const double *d_grad_out_feat,
-                                    int out_nrows, const int32_t *d_max_index,
-                                    int nchannel, cudaStream_t stream);
+template void max_pool_forward_pointer_kernel_gpu<double, int32_t, uint32_t>(
+    uint32_t *d_in_map,      //
+    uint32_t *d_out_map,     //
+    size_t const nmap,       //
+    double const *d_in_feat, //
+    double *d_out_feat,      //
+    size_t const out_nrows,  //
+    size_t const nchannel,   //
+    int32_t *d_max_index,    //
+    bool const is_sorted     //
+);
+
+template void max_pool_forward_pointer_kernel_gpu<float, int32_t, int32_t>(
+    int32_t *d_in_map,      //
+    int32_t *d_out_map,     //
+    size_t const nmap,      //
+    float const *d_in_feat, //
+    float *d_out_feat,      //
+    size_t const out_nrows, //
+    size_t const nchannel,  //
+    int32_t *d_max_index,   //
+    bool const is_sorted    //
+);
+
+template void max_pool_forward_pointer_kernel_gpu<double, int32_t, int32_t>(
+    int32_t *d_in_map,       //
+    int32_t *d_out_map,      //
+    size_t const nmap,       //
+    double const *d_in_feat, //
+    double *d_out_feat,      //
+    size_t const out_nrows,  //
+    size_t const nchannel,   //
+    int32_t *d_max_index,    //
+    bool const is_sorted     //
+);
+
+template void MaxPoolingBackwardKernelGPU<float, int32_t>(
+    float *d_grad_in_feat, size_t const in_nrows, const float *d_grad_out_feat,
+    size_t const out_nrows, const int32_t *d_max_index, size_t const nchannel);
+
+template void MaxPoolingBackwardKernelGPU<double, int32_t>(
+    double *d_grad_in_feat, size_t const in_nrows,
+    const double *d_grad_out_feat, size_t const out_nrows,
+    const int32_t *d_max_index, size_t const nchannel);
+
+// int64
+template void max_pool_forward_pointer_kernel_gpu<float, int64_t, int64_t>(
+    int64_t *d_in_map,      //
+    int64_t *d_out_map,     //
+    size_t const nmap,      //
+    float const *d_in_feat, //
+    float *d_out_feat,      //
+    size_t const out_nrows, //
+    size_t const nchannel,  //
+    int64_t *d_max_index,   //
+    bool const is_sorted    //
+);
+
+template void max_pool_forward_pointer_kernel_gpu<double, int64_t, int64_t>(
+    int64_t *d_in_map,       //
+    int64_t *d_out_map,      //
+    size_t const nmap,       //
+    double const *d_in_feat, //
+    double *d_out_feat,      //
+    size_t const out_nrows,  //
+    size_t const nchannel,   //
+    int64_t *d_max_index,    //
+    bool const is_sorted     //
+);
+
+template void MaxPoolingBackwardKernelGPU<float, int64_t>(
+    float *d_grad_in_feat, size_t const in_nrows, const float *d_grad_out_feat,
+    size_t const out_nrows, const int64_t *d_max_index, size_t const nchannel);
+
+template void MaxPoolingBackwardKernelGPU<double, int64_t>(
+    double *d_grad_in_feat, size_t const in_nrows,
+    const double *d_grad_out_feat, size_t const out_nrows,
+    const int64_t *d_max_index, size_t const nchannel);
 
 } // end namespace minkowski