+++ tests/python/tensor_field.py
@@ -68,6 +68,33 @@ class TestTensorField(unittest.TestCase):
             == {a for a in stensor.F.squeeze().detach().cpu().numpy()}
         )
 
+    def test_maxpool(self):
+        coords = torch.IntTensor(
+            [[0, 1], [0, 1], [0, 2], [0, 2], [1, 0], [1, 0], [1, 1]]
+        )
+        feats = torch.FloatTensor([[0, 1, 2, 3, 5, 6, 7]]).T
+        sfield = TensorField(feats, coords)
+
+        # Convert to a sparse tensor
+        stensor = sfield.sparse(quantization_mode=SparseTensorQuantizationMode.MAX_POOL)
+        print(stensor)
+        self.assertTrue(
+            {1, 3, 6, 7} == {a for a in stensor.F.squeeze().detach().numpy()}
+        )
+
+        # device cuda
+        if not torch.cuda.is_available():
+            return
+
+        sfield = TensorField(feats, coords, device="cuda")
+
+        # Convert to a sparse tensor
+        stensor = sfield.sparse(quantization_mode=SparseTensorQuantizationMode.MAX_POOL)
+        print(stensor)
+        self.assertTrue(
+            {1, 3, 6, 7} == {a for a in stensor.F.squeeze().detach().cpu().numpy()}
+        )
+
     def test_pcd(self):
         coords, colors, pcd = load_file("1.ply")
         voxel_size = 0.02